+++ packages/mongodb/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-mongodb
 
+## 2.1.10
+
+### Patch Changes
+
+- Updated dependencies [b3d7f59]
+- Updated dependencies [2d709ff]
+- Updated dependencies [41e8cc3]
+  - @openfn/language-common@2.3.0
+
 ## 2.1.9
 
 ### Patch Changes